requirements_complete.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
bcrypt==4.1.2  # Direct bcrypt library instead of passlib
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
alembic==1.12.1
psutil==5.9.6
pytest==7.4.3
httpx==0.25.1
redis==5.0.1
